requirements/common.txt
Django==2.2.10
django-mysql==2.4.1
djangorestframework==3.9.1
drf-nested-routers==0.91
mysqlclient==1.3.13
django-jsonfield==1.4.0
geojson==2.5.0
Shapely==1.7.0
